+++ packages/reactabular-edit/CHANGELOG.md
@@ -1,3 +1,8 @@
+1.1.0 / 2016-08-03
+==================
+
+  * Improvement - Allow `editingProps` (`value`/`onValue`) to be overridden.
+
 1.0.0 / 2016-07-25
 ==================