+++ packages/reactabular/CHANGELOG.md
@@ -1,6 +1,7 @@
 1.1.0 / 2016-08-03
 ==================
 
+  * Improvement - edit - Allow `editingProps` (`value`/`onValue`) to be overridden.
   * Improvement - table - Added `getRef` for getting references to underlying DOM elements.
   * Improvement - sticky - Added `getRef` for getting references to underlying DOM elements.
   * Improvement - sticky - Moved `reactabular-table` as a peer dependency as I realized it's better to let the user decide which version of the table to use.